requirements.txt
django==3.2.25
djangorestframework
django-cors-headers
djongo==1.3.6
djangorestframework-simplejwt
pymongo==3.12.3
cryptography
python-dotenv
